@@ -3,18 +3,19 @@
     "version": "2.0.0",
     "description": "Multitran Browser extension",
     "main": "manifest.json",
+    "type": "module",
     "scripts": {
-        "d:prepare": "esno scripts/prepare.js",
-        "js": "esno scripts/esbuild.config.js",
+        "prepare-files": "tsx scripts/prepare.js",
+        "js": "tsx scripts/esbuild.config.js",
         "tw": "tailwind -i src/common-styles.css -o extension/dist/styles.css",
         "styles:watch": "npm run tw -- -w",
         "styles:build": "npm run tw -- -m",
         "web-ext:ff": "web-ext run --source-dir extension",
         "web-ext:chrome": "web-ext run --target chromium --source-dir extension",
-        "dev:chrome": "cross-env NODE_ENV=development TARGET=chrome run-p d:prepare js styles:watch web-ext:chrome",
-        "dev:ff": "cross-env NODE_ENV=development TARGET=firefox run-p d:prepare js styles:watch web-ext:ff",
-        "build:chrome": "cross-env TARGET=chrome run-s clear d:prepare js styles:build",
-        "build:ff": "cross-env TARGET=firefox run-s clear d:prepare js styles:build",
+        "dev:chrome": "cross-env NODE_ENV=development TARGET=chrome run-p prepare-files js styles:watch web-ext:chrome",
+        "dev:ff": "cross-env NODE_ENV=development TARGET=firefox run-p prepare-files js styles:watch web-ext:ff",
+        "build:chrome": "cross-env TARGET=chrome run-s clear prepare-files js styles:build",
+        "build:ff": "cross-env TARGET=firefox run-s clear prepare-files js styles:build",
         "buildzip:chrome": "run-s build:chrome zip",
         "buildzip:ff": "run-s build:ff zip",
         "clear": "rimraf extension/dist",
@@ -27,26 +28,25 @@
     "author": "Vitaly Yerofeyevsky",
     "license": "MPL-2.0",
     "dependencies": {
-        "@babel/preset-env": "^7.19.1",
-        "autoprefixer": "^10.4.12",
+        "@babel/preset-env": "^7.22.10",
+        "autoprefixer": "^10.4.15",
         "chokidar": "^3.5.3",
         "cross-env": "^7.0.3",
-        "esbuild": "^0.15.9",
-        "eslint": "^8.23.1",
+        "esbuild": "^0.19.2",
+        "eslint": "^8.47.0",
         "eslint-config-airbnb": "^19.0.4",
-        "eslint-config-prettier": "^8.5.0",
-        "eslint-plugin-import": "^2.26.0",
-        "eslint-plugin-prettier": "^4.2.1",
-        "esno": "^0.16.3",
-        "fs-extra": "^10.1.0",
-        "jest": "^29.0.3",
-        "jest-environment-jsdom": "^29.0.3",
-        "kolorist": "^1.6.0",
+        "eslint-config-prettier": "^9.0.0",
+        "eslint-plugin-import": "^2.28.1",
+        "eslint-plugin-prettier": "^5.0.0",
+        "fs-extra": "^11.1.1",
+        "jest": "^29.6.2",
+        "jest-environment-jsdom": "^29.6.2",
+        "kolorist": "^1.8.0",
         "npm-run-all": "^4.1.5",
-        "prettier": "2.7.1",
-        "qrcode": "^1.5.1",
-        "tailwindcss": "^3.1.8",
-        "web-ext": "^7.2.0",
+        "prettier": "3.0.2",
+        "qrcode": "^1.5.3",
+        "tailwindcss": "^3.3.3",
+        "web-ext": "^7.6.2",
         "webextension-polyfill": "^0.10.0"
     },
     "jest": {
@@ -59,8 +59,9 @@
         }
     },
     "devDependencies": {
-        "@types/qrcode": "^1.5.0",
-        "@types/webextension-polyfill": "^0.9.1",
-        "rimraf": "^3.0.2"
+        "@types/qrcode": "^1.5.1",
+        "@types/webextension-polyfill": "^0.10.1",
+        "rimraf": "^5.0.1",
+        "tsx": "^3.12.7"
     }
 }
