@@ -1,5 +1,6 @@
-import { resolve } from 'path';
 import { bgCyan, black } from 'kolorist';
+import { resolve, dirname } from 'path';
+import { fileURLToPath } from 'url';
 
 export const IS_DEV = process.env.NODE_ENV === 'development';
 export const TARGET = process.env.TARGET === 'firefox' ? 'firefox' : 'chrome';
@@ -13,6 +14,9 @@ export function log(name, message) {
     // eslint-disable-next-line no-console
     console.log(black(bgCyan(` ${name} `)), message);
 }
+
+// eslint-disable-next-line no-underscore-dangle
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 /**
  *  @param  {...string} args
